Add delta compression

+++ protos/publisher_list.rb
@@ -10,7 +10,7 @@ Google::Protobuf::DescriptorPool.generated_pool.build do
       optional :compression_type, :enum, 2, "publishers_pb.PublisherList.CompressionType"
       optional :uncompressed_size, :uint32, 3
       optional :prefixes, :bytes, 4
-      repeated :deltas, :uint32, 5
+      repeated :deltas, :uint64, 5
     end
     add_enum "publishers_pb.PublisherList.CompressionType" do
       value :NO_COMPRESSION, 0
